Added textures for Vulkan TODO: Add Scene and gltf models support

+++ VoxelEngine/src/Platform/OpenGL/OpenGLBuffer.cpp
@@ -6,14 +6,14 @@
 
 namespace VoxelEngine
 {
-	OpenGLVertexBuffer::OpenGLVertexBuffer(const std::string& name, const uint32_t size)
+	OpenGLVertexBuffer::OpenGLVertexBuffer(const std::string& name, const uint64_t size)
 	{
 		glGenBuffers(1, &m_RendererID);
 		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
 		glBufferData(GL_ARRAY_BUFFER, size, nullptr, GL_STREAM_DRAW);
 	}
 
-	OpenGLVertexBuffer::OpenGLVertexBuffer(const std::string& name, const float* vertices, const uint32_t size)
+	OpenGLVertexBuffer::OpenGLVertexBuffer(const std::string& name, const float* vertices, const uint64_t size)
 	{
 		glGenBuffers(1, &m_RendererID);
 		glBindBuffer(GL_ARRAY_BUFFER, m_RendererID);
@@ -35,12 +35,12 @@ namespace VoxelEngine
 		glBindBuffer(GL_ARRAY_BUFFER, 0);
 	}
 
-	void OpenGLVertexBuffer::SetData(const void* data, const uint32_t size)
+	void OpenGLVertexBuffer::SetData(const void* data, const uint64_t size)
 	{
 		glBufferSubData(GL_ARRAY_BUFFER, 0, size, data);
 	}
 
-	OpenGLIndexBuffer::OpenGLIndexBuffer(const std::string& name, const uint32_t* indices, const uint32_t count)
+	OpenGLIndexBuffer::OpenGLIndexBuffer(const std::string& name, const uint32_t* indices, const uint64_t count)
 		: m_Count(count)
 	{
 		glGenBuffers(1, &m_RendererID);
@@ -48,7 +48,7 @@ namespace VoxelEngine
 		glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(uint32_t) , indices, GL_STATIC_DRAW);
 	}
 
-	OpenGLIndexBuffer::OpenGLIndexBuffer(const std::string& name, const uint32_t size)
+	OpenGLIndexBuffer::OpenGLIndexBuffer(const std::string& name, const uint64_t size)
 	{
 		glGenBuffers(1, &m_RendererID);
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_RendererID);
@@ -69,7 +69,7 @@ namespace VoxelEngine
 	{
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 	}
-	void OpenGLIndexBuffer::SetData(const void* data, const uint32_t size)
+	void OpenGLIndexBuffer::SetData(const void* data, const uint64_t size)
 	{
 		glBufferSubData(GL_ELEMENT_ARRAY_BUFFER, 0, size, data);
 	}
